move settingsMenuItems to setting folder
